1. Deleted "base_app". 2. Moved views from knownProgrammer into accounts app. 3. Changed way of checking if user already created programmer profile. 4. Changed variable name: existing_rating -> rating_exists
