--- knownProgrammer/base_app/admin.py
@@ -1,2 +1,0 @@
-from django.contrib import admin
-
